requirements.txt
discord.py>=2.6.3
aiohttp>=3.12.15
tiktoklive>=6.6.1
requests>=2.32.5
beautifulsoup4>=4.13.5
python-dotenv>=1.1.1
PyNaCl>=1.5.0
brotli>=1.1.0
schedule>=1.2.2
httpx>=0.25.0
openai
aiohttp>=3.12.15
beautifulsoup4>=4.13.5
brotli>=1.1.0
discord.py>=2.6.3
httpx>=0.25.0
openai
PyNaCl>=1.5.0
python-dotenv>=1.1.1
requests>=2.32.5
schedule>=1.2.2
tiktoklive>=6.6.1
